TourFer-Prototype/API_Tours/requirements.txt
Flask
Flask-MySQLdb
python-dotenv
gunicorn
mysqlclient